Refactor move things
