+++ spawn/sdk.py
@@ -7,13 +7,33 @@ from flask import Flask
 from routes.action import action_endpoint
 
 from dataclasses import (dataclass, field)
-from typing import List
+from typing import List, Callable, Any, Mapping, MutableMapping
+import inspect
 
-from entity import ActorEntity
 from internal.controller import SpawnActorController as ActorController
 
 import os
 import logging
+
+
+@dataclass
+class ActorEntity:
+
+    def command(self, name: str):
+        def register_command_handler(function):
+            """
+            Register the function to handle commands
+            """
+            # if name in self.command_handlers:
+            #    raise Exception("Command handler function {} already defined for command {}".format(
+            #        self.command_handlers[name], name))
+            # if function.__code__.co_argcount > 3:
+            #    raise Exception(
+            #        "At most three parameters, the current state, the command and the context, should be accepted by the command_handler function")
+            #self.command_handlers[name] = function
+            return function
+
+        return register_command_handler
 
 
 @dataclass
